+++ projects/evo-ui-kit/src/lib/components/evo-navigation-tabs/interfaces/evo-navigation-tab.ts
@@ -1,0 +1,12 @@
+import {ActivatedRoute, Params, QueryParamsHandling} from '@angular/router';
+
+export interface EvoNavigationTab {
+    label: string;
+    disabled?: boolean;
+    routerLink?: any[] | string | null | undefined;
+    relativeTo?: ActivatedRoute | null;
+    queryParams?: Params | null;
+    fragment?: string;
+    queryParamsHandling?: QueryParamsHandling | null;
+    preserveFragment?: boolean;
+}