+++ projects/evo-ui-kit/src/public_api.ts
@@ -68,6 +68,8 @@ export * from './lib/components/evo-upload/index';
 export * from './lib/components/evo-dropdown/index';
 export * from './lib/components/evo-navbar/index';
 export * from './lib/components/evo-navigation-button/index';
+export * from './lib/components/evo-tooltip/index';
+export * from './lib/components/evo-navigation-tabs/index';
 export * from './lib/components/evo-link-button/index';
 
 export * from './lib/pipes/index';